COMMAND, QUERY 작성 완료

+++ src/main/java/com/iiiiii/accountbook/regular_expense/query/dto/RegularExpenseDTO.java
@@ -7,12 +7,12 @@ public class RegularExpenseDTO {
     private int code;
     private int expenseDate;
     private String name;
-    private int amount;
+    private long amount;
     private int memberCode;
     private int assetCode;
     private int accCategoryCode;
 
-    public RegularExpenseDTO(int expenseDate, String name, int amount, int memberCode, int assetCode, int accCategoryCode) {
+    public RegularExpenseDTO(int expenseDate, String name, long amount, int memberCode, int assetCode, int accCategoryCode) {
         this.expenseDate = expenseDate;
         this.name = name;
         this.amount = amount;
